+++ CHANGELOG.md
@@ -3,6 +3,8 @@ HumHub Changelog
 
 1.18.0-beta.4 (Unreleased)
 ---------------------------------
+- Enh #7690: Allow uploading a custom Email Header image, or using the Logo image
+- Fix #7690: Remove "Mobile appearance" from Appearance form
 - Fix #7697: Fix submitting a form with file upload from a modal window
 - Fix #7699: Fixed old Mailer Vars usage